move js files to src folder
