+++ pages/positions/addliquidity.tsx
@@ -2,9 +2,9 @@ import { useState, useEffect, useContext, ChangeEvent, SetStateAction, Dispatch,
 import type { NextPage } from 'next'
 import Image from 'next/image'
 import { ChevronDownIcon } from '@heroicons/react/solid'
-import TokenSelectorModal from '../../components/TokenSelectorModal'
+import TokenSelectorModal from '../../src/components/TokenSelectorModal'
 import AddLiquiditySubmitButton from '../../src/components/AddLiquiditySubmitButton'
-import Tokens, { Token } from '../../components/Tokens'
+import Tokens, { Token } from '../../src/components/Tokens'
 import { AccountInfo, WalletContext } from '../../src/context/WalletContext'
 import { Provider } from '@ethersproject/providers'
 import { BigNumber, Contract } from 'ethers'
