+++ pages/positions/openloan.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next'
-import OpenLoanController from '../../components/OpenLoan/OpenLoanController'
+import OpenLoanController from '../../src/components/OpenLoan/OpenLoanController'
 
 const OpenLoanPage: NextPage = () => {
     return (
