+++ src/components/OpenLoan/OpenLoanController.tsx
@@ -3,8 +3,8 @@ import { useState, useEffect, useContext } from 'react'
 import Tokens, { Token } from '../Tokens'
 import { WalletContext } from '../../context/WalletContext'
 // TODO: import Factory from '../../abis/Factory.json'
-import PositionMgr from '../../abis/PositionManager.json'
-import IERC20 from '../../abis/ERC20.json'
+import PositionMgr from '../../../abis/PositionManager.json'
+import IERC20 from '../../../abis/ERC20.json'
 import { ethers, Contract, BigNumber, constants } from 'ethers'
 import { FieldValues } from 'react-hook-form'
 import OpenLoanView from './OpenLoanView'
